split FieldInfoSection into component, types and utils
